CLEANUP Speczilla Buffer

+++ grapher/grapher_impl.h
@@ -16,7 +16,7 @@ namespace speczilla {
 	{
 	public:
 		Buffer<T>(std::vector<T>* s);
-		explicit Buffer<T>(std::vector<T>* s, int showSamples, int samplesPerVector=1, double unitaryLength = 1, bool relative = false, double average = 0.0f, double angleScale = 1.0f);
+		explicit Buffer<T>(std::vector<T>* s, size_t showSamples, size_t samplesPerVector=1, double unitaryLength = 1, bool relative = false, double average = 0.0f, double angleScale = 1.0f);
 
 		~Buffer<T>();
 
@@ -24,7 +24,7 @@ namespace speczilla {
 
 	private:
 		std::vector<T>* buf;
-		int _samplesPerVector,_showSamples;
+		size_t _samplesPerVector,_showSamples;
 		double _unitaryLength;
 		bool _relative;
 		double _average;
